fix(api): enforce the contract error envelope and keyset code-relation paging
Close out the integration-polish items from the 2026-09-20 production
audit:

- Every /v1 handler now renders errors through ApiError, which emits the
  public-api contract envelope (code, message, retryable,
  contract_version, correlation_id + x-cortana-correlation-id header)
  instead of bare text bodies. Codes derive from the status; conflicts,
  rate limits, and server errors are marked retryable.
- Code-relation pagination anchors on the last relation id (keyset)
  instead of a numeric offset, so a concurrent re-derivation between
  pages shifts positions without skipping or repeating entries. Cursor
  payloads change shape; they are opaque and already fail closed on
  staleness.
- Name the per-page document link bound (DOCUMENT_PAGE_LINK_LIMIT).
- ContextBundle construction no longer pre-fills fields that
  with_metadata immediately finalizes.
- Consolidation attempt reports widen from u8 to u32 so pathological
  retry loops no longer saturate at 255.
- Document why the knowledge fixture's control metrics intentionally
  report graph-not-required.

+++ crates/core/src/consolidation.rs
@@ -205,7 +205,9 @@ pub struct ConsolidationOutcome {
     pub status: String,
     pub decision: ConsolidationDecisionReport,
     pub memory_id: Option<String>,
-    pub attempts: u8,
+    /// Retry count as recorded by the durable job, widened beyond u8 so the
+    /// report no longer saturates at 255 for pathological retry loops.
+    pub attempts: u32,
 }
 
 /// Evaluate a candidate without mutating a store.  Sensitive, contradictory,
@@ -371,7 +373,7 @@ pub struct QueueItem {
     pub candidate_id: String,
     pub policy_version: String,
     pub priority: u8,
-    pub attempts: u8,
+    pub attempts: u32,
     pub status: QueueStatus,
 }
 
@@ -493,7 +495,7 @@ impl ConsolidationQueue {
         None
     }
     pub fn retry(&mut self, mut item: QueueItem) {
-        item.status = if item.attempts > self.max_retries {
+        item.status = if item.attempts > u32::from(self.max_retries) {
             QueueStatus::DeadLetter
         } else {
             QueueStatus::Retry
